requirements/base.txt
Django~=2.1.0
django-environ
djangorestframework
django-storages
django-treebeard
django-imagekit==4.0.1
django-ckeditor
ftfy
django-ucamprojectlight
django-ucamwebauth
django-ucamlookup
pytz
whitenoise
Pillow
boto3
djangorestframework
markdown
django-filter
